% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semblance.R
\name{computeSemblanceOneFeature_Gini}
\alias{computeSemblanceOneFeature_Gini}
\title{Compute semblance when there is only one feature, given as a vector x, but weight the feature by its Gini coefficient. Use for data with strictly positive values.}
\usage{
computeSemblanceOneFeature_Gini(x)
}
\arguments{
\item{x}{a vector of observations for whom a given feature has been measured or estimated}
}
\value{
a Semblance metric for only one feature measured for several observations
}
\description{
Compute semblance when there is only one feature, given as a vector x, but weight the feature by its Gini coefficient. Use for data with strictly positive values.
}
