% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semblance.R
\name{makeUpperLower}
\alias{makeUpperLower}
\title{Make the upper triangular part the same as the lower triangular part.}
\usage{
makeUpperLower(m)
}
\arguments{
\item{m}{a matrix whose upper traingular part needs to be created using the lower traingular part}
}
\value{
a matrix where the upper triangular part the same as the lower triangular part
}
\description{
Make the upper triangular part the same as the lower triangular part.
}
