% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semblance.R
\name{repCol}
\alias{repCol}
\title{Make a matrix by repeating vector v into n columns}
\usage{
repCol(v, n)
}
\arguments{
\item{v}{a vector to be operated on}

\item{n}{number of columns the vector will be repeated over}
}
\value{
a matrix with repeated columns
}
\description{
Make a matrix by repeating vector v into n columns
}
