% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semblance.R
\name{repRow}
\alias{repRow}
\title{Make a matrix by repeating vector v into n rows}
\usage{
repRow(v, n)
}
\arguments{
\item{v}{a vector to be operated on}

\item{n}{number of rows the vector will be repeated over}
}
\value{
a matrix with repeated rows
}
\description{
Make a matrix by repeating vector v into n rows
}
